+++ workflow/scripts/bwtools.py
@@ -317,10 +317,10 @@ def relative_polymerase_progression(array):
 def traveling_ratio(array, res, wsize):
     return arraytools.traveling_ratio(array, wsize = wsize//res)
 
-def traveling_ratio_fixed(array, res, wsize, upstream):
-    if upstream < wsize:
-        raise ValueError("Upstream must be > %s for fixed traveling ratio. Upstream is %s"%(wsize, upstream))
-    return arraytools.traveling_ratio(array, wsize = wsize//res, peak = upstream//res)
+def traveling_ratio_fixed(array, res, wsize, relative_location):
+    if relative_location < wsize:
+        raise ValueError("Relative location (%s) must be > wsize (%s) for fixed traveling ratio."%(relative_location, wsize))
+    return arraytools.traveling_ratio(array, wsize = wsize//res, peak = relative_location//res)
 
 def summit_loc(array, res, wsize, upstream):
     loc = arraytools.relative_summit_loc(array, wsize = wsize//res)
@@ -391,9 +391,9 @@ def query_main(args):
             'max' : np.nanmax,
             'min' : np.nanmin,
             'RPP' : relative_polymerase_progression,
-            'TR' : lambda array: traveling_ratio(array, args.res, 50),
-            'TR_fixed' : lambda array: traveling_ratio_fixed(array, args.res, 50, args.upstream),
-            'summit_loc': lambda array: summit_loc(array, args.res, 50, args.upstream)}
+            'TR' : lambda array: traveling_ratio(array, args.res, args.wsize),
+            'TR_fixed' : lambda array: traveling_ratio_fixed(array, args.res, args.wsize, args.upstream + args.TR_A_center),
+            'summit_loc': lambda array: summit_loc(array, args.res, args.wsize, args.upstream)}
     try:
         summary_func = summary_funcs[args.summary_func]
     except KeyError:
@@ -635,6 +635,8 @@ if __name__ == "__main__":
             'identity' summary. mean, median, max, min supported. Additionally, traveling ratio ('TR') and relative polymerase progression ('RPP'), \
             and 'summit_loc' (local peak identification) are supported. Default = 'mean'", default = "mean")
     parser_query.add_argument('--gzip', action = "store_true", help = "gzips the output if flag is included")
+    parser_query.add_argument('--TR_A_center', type = int, help = "center of window A in fixed traveling ratio. In relative bp to region start")
+    parser_query.add_argument('--wsize', type = int, default = 50, help = "Size of half window in bp for calcs that use windows. Default = 50 bp")
     parser_query.set_defaults(func=query_main)
 
     # compare verb